Updating template generator
