+++ example/templates/default/bin/schema.dart
@@ -1,7 +1,7 @@
 import 'package:wildfire/wildfire.dart';
 
 main() {
-  var dataModel = new DataModel(WildfireSink.modelTypes());
+  var dataModel = new DataModel.fromPackageContainingType(WildfireSink);
   var persistentStore = new PostgreSQLPersistentStore(() => null);
   var ctx = new ModelContext(dataModel, persistentStore);
