+++ example/templates/default/pubspec.yaml
@@ -8,7 +8,8 @@ environment:
 
 dependencies:
   scribe: any
-  aqueduct: any
+  aqueduct:
+    path: /Users/joeconway/Projects/aqueduct
 
 dev_dependencies:
   test: '>=0.12.0 <0.13.0'
